load task from file
